fix: добавляет объект с фильтрами, переносит функцию getPhotosToRender в файл фильтров
